+++ frontend/src/components/TripsSection/TripsList/Trip/Trip.module.css
@@ -1,0 +1,16 @@
+.trip {
+  display: inline-block;
+  width: 250px;
+  height: 100px;
+  margin: 10px;
+}
+
+.trip Button {
+  width: 100%;
+  height: 100%;
+}
+
+.trip Button:hover {
+  font-weight: bold;
+  background-color: #a8d0e6;
+}
