+++ frontend/src/components/TripsSection/TripsSection.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import styles from './TripsSection.module.css';
-import Spinner from '../../commons/Spinner/Spinner';
-import TripsList from '../commons/TripsList/TripsList';
+import Spinner from '../commons/Spinner/Spinner';
+import TripsList from './TripsList/TripsList';
 
-export default function GroupsSection({ title, trips, loading, disable }) {
+export default function TripsSection({ title, trips, loading, disable }) {
   return (
     <div className={styles.section}>
       <h1>{title}</h1>
