+++ frontend/src/components/TripsSection/TripsSection.module.css
