+++ frontend/src/pages/DashboardPage/DashboardPage.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styles from './DashboardPage.module.css';
 import HomeTopBar from '../../components/HomeTopBar/HomeTopBar';
-import TripsSection from '../../components/dashboard/TripsSection/TripsSection';
+import TripsSection from '../../components/TripsSection/TripsSection';
 
 import useGet from '../../hooks/useGet';
